shortform-maker/requirements.txt
streamlit>=1.40.0
openai>=1.50.0
anthropic>=0.40.0
yt-dlp>=2024.12.0
python-dotenv>=1.0.0
